LayoutTests/http/tests/security/mixedContent/about-blank-iframe-in-main-frame.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.waitUntilDone();
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
    layoutTestController.setCanOpenWindows();
    layoutTestController.setCloseRemainingWindowsWhenComplete(true);
}

window.addEventListener("message", function (e) {
  if (window.layoutTestController)
      layoutTestController.notifyDone();
}, false);

</script>
<p>This test opens a window iframe that loads an about:blank iframe.  We should
*not* trigger a mixed content callback because the about:blank iframes cannot
be corrupted by active network attackers.</p>
<script>
window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-about-blank-frame.html");
</script>
</body>
</html>
