Safari-609.1.11.1/LayoutTests/http/tests/security/mixedContent/about-blank-iframe-in-main-frame-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - didCommitLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
frame "<!--frame1-->" - didCommitLoadForFrame
frame "<!--frame1-->" - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFinishLoadForFrame
frame "<!--frame2-->" - didStartProvisionalLoadForFrame
frame "<!--frame2-->" - didCommitLoadForFrame
frame "<!--frame2-->" - didFinishDocumentLoadForFrame
frame "<!--frame2-->" - didHandleOnloadEventsForFrame
frame "<!--frame2-->" - didFinishLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This test opens a window iframe that loads an about:blank iframe. We should *not* trigger a mixed content callback because the about:blank iframes cannot be corrupted by active network attackers.
